os unlink in dummy _init

+++ mlens/utils/dummy.py
@@ -687,25 +687,18 @@ def _init(train, label, shape):
     """Simple temp folder initialization for testing estimation functions."""
 
     dir = os.path.join(os.getcwd(), 'tmp')
+    if os.path.exists(dir):
+        os.unlink(dir)
+        try:
+            shutil.rmtree(dir)
+        except OSError:
+            # Can fail on windows
+            dlc = subprocess.Popen('rmdir /S /Q %s' % dir,
+                                   shell=True,
+                                   stdout=subprocess.PIPE,
+                                   stderr=subprocess.PIPE)
 
-    # Some fail safes before setting up cache
-    try:
-        # Remove dir if already exists
-        shutil.rmtree(dir)
-    except:
-        pass
-
-    try:
-        # Shutil can fail to remove the dir on windows,
-        # if mkdir fails we do a force removal and try again.
-        os.mkdir(dir)
-    except:
-        # Force remove and try again
-        dlc = subprocess.Popen('rmdir /S /Q %s' % dir,
-                               shell=True,
-                               stdout=subprocess.PIPE,
-                               stderr=subprocess.PIPE)
-        os.mkdir(dir)
+    os.mkdir(dir)
 
     paths = {}
     for name, arr in zip(('X', 'y'), (train, label)):
